+++ lib/core/theme/app_theme.dart
@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:malawi_ride_share_app/core/theme/app_bar_theme.dart';
-import 'package:malawi_ride_share_app/core/theme/app_colors.dart';
-import 'package:malawi_ride_share_app/core/theme/app_elevation.dart';
-import 'package:malawi_ride_share_app/core/theme/app_radii.dart';
-import 'package:malawi_ride_share_app/core/theme/app_spacing.dart';
-import 'package:malawi_ride_share_app/core/theme/app_typography.dart';
-import 'package:malawi_ride_share_app/core/theme/app_typography_token.dart';
-import 'package:malawi_ride_share_app/core/theme/bottom_sheet_theme.dart';
-import 'package:malawi_ride_share_app/core/theme/button_themes.dart';
-import 'package:malawi_ride_share_app/core/theme/dialog_theme.dart';
-import 'package:malawi_ride_share_app/core/theme/input_decoration_theme.dart';
-import 'package:malawi_ride_share_app/core/theme/snack_bar_theme.dart';
+import 'package:malawi_ride_share_app/core/theme_components/app_bar_theme.dart';
+import 'package:malawi_ride_share_app/core/tokens/app_colors.dart';
+import 'package:malawi_ride_share_app/core/tokens/app_elevation.dart';
+import 'package:malawi_ride_share_app/core/tokens/app_radii.dart';
+import 'package:malawi_ride_share_app/core/tokens/app_spacing.dart';
+import 'package:malawi_ride_share_app/core/theme_components/app_typography.dart';
+import 'package:malawi_ride_share_app/core/tokens/app_typography_token.dart';
+import 'package:malawi_ride_share_app/core/theme_components/bottom_sheet_theme.dart';
+import 'package:malawi_ride_share_app/core/theme_components/button_themes.dart';
+import 'package:malawi_ride_share_app/core/theme_components/dialog_theme.dart';
+import 'package:malawi_ride_share_app/core/theme_components/input_decoration_theme.dart';
+import 'package:malawi_ride_share_app/core/theme_components/snack_bar_theme.dart';
 
 ThemeData buildAppTheme() {
   return ThemeData(
